Switch to sketchy manual blin filtering

+++ libs/wgpu-buffer/terrain/shaders/accumulate_spherical_normals.comp.glsl
@@ -36,7 +36,7 @@ main()
         // Load the relevant normal sample.
         vec2 grat = texelFetch(sampler2D(terrain_deferred_texture, terrain_linear_sampler), coord, 0).xy;
         uint atlas_slot = terrain_atlas_slot_for_graticule(grat, index_texture, index_sampler);
-        ivec2 raw_normal = terrain_normal_in_tile(grat, tile_info[atlas_slot], atlas_texture, atlas_sampler);
+        vec2 raw_normal = terrain_normal_in_tile(grat, tile_info[atlas_slot], atlas_texture, atlas_sampler);
 
         // FIXME: blend normal with existing buffer.
